requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-configurations==2.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.8.6
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
Werkzeug==1.0.1
whitenoise==5.2.0
